Updated naming conventions, removed unndcecary imports and refactored components
