+++ client/src/App.js
@@ -1,12 +1,11 @@
-import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { ApolloClient, ApolloProvider, InMemoryCache } from "@apollo/client";
 import Dashboard from "./pages/Dashboard";
-import Navbar from "./components/nav";
+import Navbar from "./components/Nav";
 import NotFound from "./pages/NotFound";
-import Homepage from "./pages/homepage";
-import Login from "./pages/login";
-import Signup from "./pages/signup";
+import Homepage from "./pages/Homepage";
+import Login from "./pages/Login";
+import Signup from "./pages/Signup";
 
 const client = new ApolloClient({
   uri: "/graphql",
