--- client/src/components/viewGradesPage.js
@@ -1,20 +1,0 @@
-export default function gradesPage() {
-    return (
-        <div>
-
-            <div
-                class="image"
-                style={{
-                    height: "350px",
-                    width: "550px",
-                    backgroundImage:
-                        'url("client\src\Assets\happyBalloons.jpg")',
-                    backgroundSize: "contain",
-                    backgroundRepeat: "no-repeat",
-                }}
-            >
-                This div contains the grades page background image.
-            </div>
-        </div>
-    );
-}
