+++ client/src/pages/Dashboard.jsx
@@ -1,6 +1,5 @@
-import React from 'react'
 import styles from './dashboard.module.css'
-import { Outlet, Link } from "react-router-dom";
+import { Link } from "react-router-dom";
 
 const dummyData = {
   name: "John Doe",
