+++ client/src/pages/NotFound.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useLocation } from 'react-router-dom';
 
 function NotFound() {
